+++ mycollab-services/src/main/java/com/esofthead/mycollab/module/ecm/service/DriveInfoService.java
@@ -15,5 +15,5 @@ public interface DriveInfoService extends ICrudService<Integer, DriveInfo> {
 	DriveInfo getDriveInfo(@CacheKey Integer sAccountId);
 
 	@Cacheable
-	Double getUsedStorageVolume(@CacheKey Integer sAccountId);
+	Long getUsedStorageVolume(@CacheKey Integer sAccountId);
 }
